+++ Craftsman/Builders/ValueObjectBuilder.cs
@@ -42,13 +42,14 @@ public static class ValueObjectBuilder
     // source: https://github.com/jhewlett/ValueObject
     public abstract class ValueObject : IEquatable<ValueObject>
     {{
-	    private List<PropertyInfo>? properties;
-	    private List<FieldInfo>? fields;
+        private List<PropertyInfo>? _properties;
+        private List<FieldInfo>? _fields;
+
 	    public static bool operator ==(ValueObject? obj1, ValueObject? obj2)
 	    {{
-		    if (object.Equals(obj1, null))
+		    if (Equals(obj1, null))
 		    {{
-			    if (object.Equals(obj2, null))
+			    if (Equals(obj2, null))
 				    return true;
 
 			    return false;
@@ -76,27 +77,27 @@ public static class ValueObjectBuilder
 
 	    private bool PropertiesAreEqual(object obj, PropertyInfo p)
 	    {{
-		    return object.Equals(p.GetValue(this, null), p.GetValue(obj, null));
+		    return Equals(p.GetValue(this, null), p.GetValue(obj, null));
 	    }}
 
 	    private bool FieldsAreEqual(object obj, FieldInfo f)
 	    {{
-		    return object.Equals(f.GetValue(this), f.GetValue(obj));
+		    return Equals(f.GetValue(this), f.GetValue(obj));
 	    }}
 
 	    private IEnumerable<PropertyInfo> GetProperties()
-	    {{		
-            return this.properties ?? (this.properties = GetType()
-                .GetProperties(BindingFlags.Instance | BindingFlags.Public)
-                .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
-                .ToList());
+	    {{
+            return _properties ??= GetType()
+	            .GetProperties(BindingFlags.Instance | BindingFlags.Public)
+	            .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
+	            .ToList();
 	    }}
 
 	    private IEnumerable<FieldInfo> GetFields()
 	    {{
-            return this.fields ?? (this.fields = GetType().GetFields(BindingFlags.Instance | BindingFlags.Public)
-                .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
-                .ToList());
+            return _fields ??= GetType().GetFields(BindingFlags.Instance | BindingFlags.Public)
+	            .Where(p => p.GetCustomAttribute(typeof(IgnoreMemberAttribute)) == null)
+	            .ToList();
 	    }}
 
 	    public override int GetHashCode()